rotas autenticadas + regra de negócio com o usuário logado

+++ server/src/routes/memories.ts
@@ -4,9 +4,18 @@ import { z } from 'zod'
 
 
 export async function memoriesRoutes(app: FastifyInstance){
+
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify()
+    })
     
-    app.get('/memories', async () => {
+    app.get('/memories', async (request) => {
+
+        
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub
+            },
             orderBy: {
                 createdAt: 'asc',
             }
@@ -20,7 +29,7 @@ export async function memoriesRoutes(app: FastifyInstance){
         })
     })
 
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, reply) => {
         //const { id } = request.params
 
         const paramsSchema = z.object({
@@ -34,6 +43,10 @@ export async function memoriesRoutes(app: FastifyInstance){
                 id,
             }
         })
+
+        if(!memory.isPublic && memory.userId != request.user.sub){
+            return reply.status(401).send()
+        }
         return memory
     })
 
@@ -51,18 +64,28 @@ export async function memoriesRoutes(app: FastifyInstance){
                 content, 
                 coverUrl, 
                 isPublic,
-                userId: 'f02aae2c-5eb4-45bb-8b61-24db3c2fbbe9',
+                userId: request.user.sub,
             }
         })
         return memory
     })
 
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string(),
         })
 
         const { id } = paramsSchema.parse(request.params)
+
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if(memory?.userId != request.user.sub){
+            return reply.status(401).send()
+        }
 
         const bodySchema = z.object({
             content: z.string(),
@@ -86,12 +109,22 @@ export async function memoriesRoutes(app: FastifyInstance){
         
     })
 
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
 
         const { id } = paramsSchema.parse(request.params)
+
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if(memory?.userId != request.user.sub){
+            return reply.status(401).send()
+        }
 
         await prisma.memory.delete({
             where: {
